requirements.txt
# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=4.9.3

# Environment variable management
python-dotenv>=1.0.0

#for playstore reviews
google-play-scraper